+++ uui/src/lib/List/ListGroup.types.d.ts
@@ -1,1 +1,3 @@
-export type ListGroupProps = svelte.JSX.HTMLAttributes<HTMLDivElement>;
+import type { PaperProps } from '../Paper/Paper.types';
+
+export type ListGroupProps<T extends EventTarget> = PaperProps<T>;