+++ website/src/routes/list-group/list-group-exp.svelte
@@ -1,22 +1,20 @@
 <script lang="ts">
-  import { ListGroup, List, ListItem, ListItemAdornment, ListSubHeader } from '@uui';
+  import { ListGroup, List, ListItem, ListItemAdornment, ListSubHeader, Typography } from '@uui';
 </script>
 
-<ListGroup class="demo-list">
+<ListGroup class="demo-list-group">
   <ListSubHeader>List 1</ListSubHeader>
   <List>
-    <ListItem text="List Item" divider>
-      <ListItemAdornment class="material-symbols-outlined" slot="leading-adornment">
-        star
-      </ListItemAdornment>
+    <ListItem divider>
+      <Typography>List Item</Typography>
+      <ListItemAdornment class="material-symbols-outlined" slot="leading">star</ListItemAdornment>
     </ListItem>
   </List>
   <ListSubHeader>List 2</ListSubHeader>
   <List>
-    <ListItem text="List Item">
-      <ListItemAdornment class="material-symbols-outlined" slot="leading-adornment">
-        star
-      </ListItemAdornment>
+    <ListItem>
+      <Typography>List Item</Typography>
+      <ListItemAdornment class="material-symbols-outlined" slot="leading">star</ListItemAdornment>
     </ListItem>
   </List>
 </ListGroup>