+++ website/src/routes/menu/menu-on-clickaway.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { Menu, Button, ListItem, Box } from '@uui';
+  import { Menu, Button, ListItem, Box, Typography } from '@uui';
 
   let anchorRef: HTMLElement;
   let open = false;
@@ -13,6 +13,6 @@
 <Box ssx={{ $self: { display: 'flex', gap: '5rem', 'justify-content': 'center' } }}>
   <Button variant="text" on:click={() => (open = !open)} bind:ref={anchorRef}>Clickaway</Button>
   <Menu {open} bind:anchorRef anchorOrigin="bottom-start" onClickaway={clickaway}>
-    <ListItem text="List Item" />
+    <ListItem><Typography>List Item</Typography></ListItem>
   </Menu>
 </Box>