+++ website/src/routes/menu/menu-top-start.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { Menu, Button, ListItem, ListItemAdornment, Box } from '@uui';
+  import { Menu, Button, ListItem, ListItemAdornment, Box, Typography } from '@uui';
 
   let open = false;
   let anchorRef: HTMLElement;
@@ -8,19 +8,17 @@
 <Box ssx={{ $self: { display: 'flex', gap: '5rem', 'justify-content': 'center' } }}>
   <Button variant="text" on:click={() => (open = !open)} bind:ref={anchorRef}>top-start</Button>
   <Menu bind:open bind:anchorRef anchorOrigin="top-start">
-    <ListItem text="List Item">
-      <ListItemAdornment slot="leading-adornment" class="material-symbols-outlined">
-        send
-      </ListItemAdornment>
-      <ListItemAdornment slot="trailing-adornment" class="material-symbols-outlined">
+    <ListItem>
+      <Typography>List Item</Typography>
+      <ListItemAdornment slot="leading" class="material-symbols-outlined">send</ListItemAdornment>
+      <ListItemAdornment slot="trailing" class="material-symbols-outlined">
         expand_more
       </ListItemAdornment>
     </ListItem>
-    <ListItem text="List Item">
-      <ListItemAdornment slot="leading-adornment" class="material-symbols-outlined">
-        mail
-      </ListItemAdornment>
-      <ListItemAdornment slot="trailing-adornment" class="material-symbols-outlined">
+    <ListItem>
+      <Typography>List Item</Typography>
+      <ListItemAdornment slot="leading" class="material-symbols-outlined">mail</ListItemAdornment>
+      <ListItemAdornment slot="trailing" class="material-symbols-outlined">
         expand_more
       </ListItemAdornment>
     </ListItem>